requirements.txt
jsonpath-ng==1.7.0
requests==2.32.5
